+++ src/module.js
@@ -7,8 +7,7 @@ import {
   applySystemSpecificStyles,
   checkSystem,
   patchCurrencySettings,
-  registerSettings,
-  applyShims
+  registerSettings
 } from "./settings.js";
 import { registerHotkeysPost, registerHotkeysPre } from "./hotkeys.js";
 import Socket from "./socket.js";
@@ -32,7 +31,6 @@ Hooks.once("init", async () => {
   registerLibwrappers();
   registerUIOverrides();
   setupCaches();
-  applyShims();
   setupPlugins("init");
 });