+++ changelog.md
@@ -3,6 +3,7 @@
 ## Version 2.6.17
 
 - Added services to the "Populate Items" tab in merchants, and buttons to clear them
+- Fixed PF2e mystified items incorrectly showing the item's real name and icon in the user-to-user trading interface instead of the mystified ones
 - Fixed item piles not being deleted once empty
 - Fixed item piles being updated with `actorData` when that has been deprecated in Foundry V11