+++ changelog.md
@@ -8,6 +8,7 @@
 - Fixed item piles being updated with `actorData` when that has been deprecated in Foundry V11
 - Fixed price preset editor not opening
 - Fixed not being able to turn actors into item piles through the actor sheet header
+- Fixed item piles having problems with the new Foundry V11 synthetic actor methodology in some systems
 
 ## Version 2.6.16