+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Version 2.6.17
 
+- Updated Polish localization (thank you Lioheart on Weblate!)
 - Added services to the "Populate Items" tab in merchants, and buttons to clear them
 - Fixed PF2e mystified items incorrectly showing the item's real name and icon in the user-to-user trading interface instead of the mystified ones
 - Fixed item piles not being deleted once empty