+++ changelog.md
@@ -1,8 +1,9 @@
 # Item Piles Changelog
 
-## Version 2.6.17
+## Version 2.7.0
 
 - Updated Polish localization (thank you Lioheart on Weblate!)
+- Added a "Restrict Access" setting and "View" permissions to vault access configuration - with these two configured, only characters with the permissions set up can view the contents of the vault
 - Added services to the "Populate Items" tab in merchants, and buttons to clear them
 - Added an export button to the settings UI to export and import the system specific settings
 - Fixed PF2e mystified items incorrectly showing the item's real name and icon in the user-to-user trading interface instead of the mystified ones